feature(TDQ-13885) : code review

+++ dataquality-semantic/src/main/java/org/talend/dataquality/semantic/recognizer/CategoryFrequency.java
@@ -32,7 +32,7 @@ public class CategoryFrequency implements Comparable<CategoryFrequency>, Seriali
 
     int categoryLevel;
 
-    float score;
+    double score;
 
     /**
      * CategoryFrequency constructor from a category.
@@ -94,11 +94,11 @@ public class CategoryFrequency implements Comparable<CategoryFrequency>, Seriali
     }
 
     /**
-     * @deprecated
+     * @deprecated use getScore() instead
      * @return
      */
     public float getFrequency() {
-        return score;
+        return (float) score;
     }
 
     public long getCount() {
@@ -109,7 +109,7 @@ public class CategoryFrequency implements Comparable<CategoryFrequency>, Seriali
         return categoryLevel;
     }
 
-    public float getScore() {
+    public double getScore() {
         return score;
     }
 
@@ -140,7 +140,7 @@ public class CategoryFrequency implements Comparable<CategoryFrequency>, Seriali
     public int compareTo(CategoryFrequency o) {
         // The EMPTY category must always be ranked after the others
 
-        int count = Float.compare(this.getScore(), o.getScore());
+        int count = Double.compare(this.getScore(), o.getScore());
         if (count != 0) {
             return count;
         } else {
